moved TestOrderProvider so it is now available from project-base
